Fix strange crash when looking at exits.
Not sure when this bug was introduced, but apparently it's been there a while. The bug happens on line 91 or line 102 depending on god knows what. Most perplexing at all is that items can be looked at fine, but when calling look on
an exit, the item code which runs to search for item matches before the exit search code even runs is what crashes! This makes no sense to me currently, so I'm just adding a check for None value which should have been there anyway,
and we can call this confusing but fixed.