Fix #113.
Check rooms for offline users every time a user loads the room, not just the
first time since engine launch. Probably the users are not being removed
from the room occupant lists because they are uncleanly losing
connection. Checking for and removing users should not be a very
expensive operation, so it is fine to do it every time.